requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
docopt==0.6.2
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
pillow==11.3.0
PyYAML==6.0.3
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
rpds-py==0.27.1
svgwrite==1.4.3
urllib3==2.5.0
Werkzeug==3.1.3
yarg==0.1.10
